requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
psycopg2-binary
python-dotenv==1.1.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
